Makes dictionary of names recorded in year/decade
Can now tie years with their names recorded in that given year. These results can then be merged such that names recorded in a given decade can be found. Sorting can be applied to find the most popular names recorded at birth in a given year/decade. Plotting this data is still an issue as the text isn't showing on the bars and there may still be some debugging necessary as some names don't seem to be correlating to the correct bar - this only seems to be an issue in plotting though; reading the values directly from an outputted dictionary shows they seem to be correct intially.