Reorganised folder
Functions related to manipulating dictionaries are in their own file. 
The main file imports this and namesnlp.py to use their functions, while still containing the threading and plotting code and lines demonstrating currying the functions to get data. 
A basic function for finding close variations of a name in the set of names has been written but is slow since the Jaro-Winkler distance function I've written is not fast for names with a low distance, i.e. Marie and Eerie seem to be very slow compared to Jake and Jacob or Mary and Marie, hence this will need to be rewritten to be faster, potentially through utilising parallelism or just simplifying the code and requiring the user to input data differently.